refactoring and add custom errors
for better errors wanted to have the file path in the decompiler but
std::fs::File doesn't keep that so reworked to pass the PathBuf and have the
decompiler open the file. Then also refactored to make most of the methods utility methods and pass in the buffer to read from.
Then added a custom Result type alias using the new error enum which leverages
thiserror crate.